code refacto
