+++ custom-recipes/image-processing-custom/recipe.py
@@ -3,7 +3,8 @@ from PIL import Image
 from io import BytesIO
 import numpy as np
 import logging
-from ocr_utils import get_input_output, image_processing_parameters
+from ocr_recipes_io_utils import get_input_output
+from ocr_utils import image_processing_parameters
 from ocr_constants import Constants
 
 logger = logging.getLogger(__name__)
